Upgrade to Cake 3.10
