+++ bin/Installer.php
