+++ composer.json
@@ -6,13 +6,13 @@
 	"license" : "MIT",
 	"require" : {
 		"php" : ">=7.4",
-		"cakephp/cakephp" : ">=3.6 <3.7.0",
+		"cakephp/cakephp" : "^3.10",
 		"vlucas/phpdotenv" : "~3.6.8",
 		"mobiledetect/mobiledetectlib" : "2.*",
-		"cakephp/migrations" : "~1.0",
+		"cakephp/migrations" : "*",
 		"cakephp/plugin-installer" : "*",
 		"friendsofcake/bootstrap-ui" : "^1.4",
-		"friendsofcake/cakephp-csvview" : "~3.0",
+		"friendsofcake/cakephp-csvview" : "^3.0",
 		"friendsofcake/crud" : "^5.0",
 		"josegonzalez/cakephp-upload" : "^4.0",
 		"braintree/braintree_php" : "~3.0",
@@ -27,9 +27,8 @@
 		"directorytree/ldaprecord": "2.*"
 	},
 	"require-dev" : {
-		"psy/psysh" : "@stable",
-		"cakephp/debug_kit" : "~3.2",
-		"cakephp/bake" : "~1.1"
+		"cakephp/debug_kit" : "^3.23",
+		"cakephp/bake" : "*"
     },
 	"suggest" : {
 		"phpunit/phpunit" : "Allows automated tests to be run without system-wide install.",
