+++ src/Controller/EventsControllerTest.php
@@ -16,16 +16,16 @@ class EventsControllerTest extends IntegrationTestCase
      * @var array
      */
     public $fixtures = [
-        'app.events',
-        'app.rooms',
-        'app.contacts',
-        'app.prerequisites',
-        'app.honoraria',
-        'app.categories',
-        'app.tools',
-        'app.files',
-        'app.registrations'
-    ];
+		'app.events',
+		'app.rooms',
+		'app.contacts',
+		'app.prerequisites',
+		'app.honorarias',
+		'app.categories',
+		'app.tools',
+		'app.files',
+		'app.registrations'
+	];
 
     /**
      * Test beforeFilter method
