--- tests/TestCase/Controller/CategoriesControllerTest.php
@@ -1,52 +1,0 @@
-<?php
-namespace App\Test\TestCase\Controller;
-
-use App\Controller\CategoriesController;
-use Cake\TestSuite\IntegrationTestCase;
-
-/**
- * App\Controller\CategoriesController Test Case
- */
-class CategoriesControllerTest extends IntegrationTestCase
-{
-
-    /**
-     * Fixtures
-     *
-     * @var array
-     */
-    public $fixtures = [
-        'app.categories',
-        'app.events'
-    ];
-
-    /**
-     * Test beforeFilter method
-     *
-     * @return void
-     */
-    public function testBeforeFilter()
-    {
-        $this->markTestIncomplete('Not implemented yet.');
-    }
-
-    /**
-     * Test isAuthorized method
-     *
-     * @return void
-     */
-    public function testIsAuthorized()
-    {
-        $this->markTestIncomplete('Not implemented yet.');
-    }
-
-    /**
-     * Test index method
-     *
-     * @return void
-     */
-    public function testIndex()
-    {
-        $this->markTestIncomplete('Not implemented yet.');
-    }
-}
