+++ webroot/css/app.css
@@ -1,7 +1,7 @@
 body {
   font-family: "Source Sans Pro", Calibri, Candara, Arial, sans-serif;
   font-size: 15px;
-  margin-top: 80px;
+  margin-top: 0px;
   margin-bottom: 30px;
   -moz-osx-font-smoothing: grayscale;
   -webkit-font-smoothing: antialiased;
